scraper/requirements.txt
lxml
playwright
python-dateutil
beautifulsoup4
markdown
langchain-community